Fix even more dependency cycles (#6243)
Refs #6177
